fix: Color and warmth changes to fix bugs, extra colors
There was a bug in how I was assigning color and warmth.

Color: I always assumed brightness 100 (if not mentioned, it would default
to 100) which heavily effects the actual color output. For some cases, I'm
allowing special brightness and warmth settings. I also fixed a typo that
allowed a brightness to be set too high.

White: Defaults to 3500 as per rebecca's preference.

Extra colors: Fixed the colors I had ( they were way off) and added
orange + purple as fan favs

Warmth changes:
Warmth was assumed prior to be from 0 to 3500, which I now realize I just
misread the docs. Warmth is actually from 2500 to 9000 with a default of
3500. As such, I changed the min/max settings and also the interval of
change.